Fix upgrades
